Splits file into Setup, Definitions and Functions sections.
Locates the project root of the script, so that the working directory can be external to the project directory.

Puts pipeline and dry_run arguments into parsers, and leaves space for kwargs.

If no pipeline is given, will try to run the Main pipeline (which does not currently do anything).